requirements-dev.txt
pytest==8.1.1
pytest-cov==4.1.0
pandas==2.2.1
pyspark==3.5.1
pydantic==2.6.4
PyYAML==6.0.1 